Update templeated repo.
This change replaces some of the boilerplate
generated in the initial commit.

This change removes the use of pytest for test authoring
but keeps it for test running.

This change updates tox and pre-commit we a basic set of
default checks and test envionments.

This change removes the use of black for linting, flake8
and autopep8 will be used instead.

This change adds a stub arbiterd-static console script