prepare basic test framework and add inital filesystem support
This change renames the test direcotry to arbiterd_tests
to ensure a unique name when importing the test package
in the unit tests.

This change starts developing the common filesystem module
to ensure the testing apporch will work.

This change intoduces a fake sysfs file system tree in
arbiterd_tests/test_data/sys and asserts its importable
and discoverable in the unit tests.

setup.cfg and tox.ini are modifed to ensure the test data is not
included in the package and that pytest can discover the test
without recursing into the fake sys tree.