Add ATTestCase and fix tox and setup.cfg configs.
This change creates a base test class for arbiterd
that provides a mocked /sys filesystem and mocked logging.

This change also corrects the tox.ini envs to reflect
that tests was renamed to arbiterd_tests a previous commit.

testtools is now listed as a testing dep in setup.cfg
testtools extends the stanard lib unittesting package
with supprot for fixtures, better asserts and other enhancements.
testtools.TestCase is therefor used as the base of the new ATTestCase
class to inherit those enhancements.