Add etc fixture and basic cli
This change adds a nova.conf sample file and ETC functional
test fixture. This will enable future work to parse the nova.conf
and extract the range of cpus managed by nova.

This change add some basic cli commands to list the avaiable,
online and offline cpus and adds TODOs for other work in the future.