Add inital tooling to automate packaging
This change introduces tools/arbiter-ctl.sh
to automate building arbiterd as a contaienr.

it will be extneded in the future to manage installation
of arbiterd as a systemd service and its config files.

This change also modifes tox to build only the binary wheel as part
of the build env and adds a sdist env to build the sdist package.