add cpu-state arbiter
This change adds the inital implementation of the cpu-state arbiter.

This change extend the static cli with two new commands arbitrate
and revoke which allow the cpu-state arbiter to be invoked.
Both commands default to dry-run mode which for the cpu-state
arbiter prints the cores that would be offlined or onlined.
To make the change take effect --apply must be passed.

This change also introduces the dispatcher class which is responciple
for loading the arbiters and dispatching the request to them.

finally this change adds context, hardware_thread and instance objects
to model the respective concepts.

The objects module contains the high level objects that model the
system. the object typically will delegate the logic of perfroming
oeprations on the system to lib modules stored in arbiter.common