[WIP] add third party tcib image.
This change try to follow the third party container
image creation docs for tripleo to care an arbiterd
container image that extends the existing nova-libvirt container.
nova libvirt is currenlty used to ensure libvirt is present to allow
libvirt-python to install and to ensure the nova user and group
exits with the correct uid and gid.

Arbiterd runs as the arbiterd user which is also a member of
the libvirt group and nova group to ensure we can read novas
config file and connect to the libvirt deamon.
The container will install arbiterd from pypi in a python 3.8
virtual enviornment.
/opt/arbiterd/etc/nova is created in the image.
when runing the image the nova config directory should be
bind mount readonly to this localtion so arbiterd can access
teh nova.conf.

arbiterd need elevated permissions to function and
/sys/devices/system/bus/cpu need to be bind mount rw
from the host at a minium to enable arbiterd to manage the online state
/run will also need to be shared with the nova-libvirt contaienr
to enabled acess to the libvirtd sockets.