fix apply and hashing.
as part of the unit test work the HardwareThread
object were reused when caluating the cores to offline
instead of workign with sets fo ints.

since the HardwareThread class did not define a hash fucntion they
are not valid member of a set. this is not seen in the unittests
since the autospec mock has a hash function.

this change also fixs the --apply defintion to be a bool not
an int.